fix(ci): file-line-limit — split companion-env tests into part_016
test_executor_error_paths_part_006.py was 1504 lines, over the
1500-line first-party maintainability limit. Move the self-contained
companion-env secret-refresh + compose-environment test group (which
only depends on monitor_handoff / companion_services, not the executor
scaffolding) into a new test_executor_error_paths_part_016.py and drop
the now-unused imports (yaml, companion_services, executor_monitor_handoff,
monitor_handoff_companion_env) from part_006. part_006 is now 1062 lines
and part_016 is 458; all moved tests still pass.
